requirements.txt
networkx
pandas
python-louvain
openai
python-dotenv
plotly
scikit-learn
pymysql
tqdm
